Restructure + Packet Handling
Restructured the VITA backend again. Got a fair amount of the packet handling finished, currently incorporating Travis' existing vrt_command.c code
