+++ iiod/vita49_2/vita49_2_host.h
@@ -1,0 +1,26 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+/*
+ * libiio - Library for interfacing industrial I/O (IIO) devices
+ *
+ * Copyright (C) 2026 Analog Devices, Inc.
+ * Author: Travis Collins <travis.collins@analog.com>
+ * 
+ * Contributors:
+ * 		- Praveen Perera <praveen.perera@analog.com>
+ */
+
+#ifndef __VITA49_2_BACKEND_H__
+#define __VITA49_2_BACKEND_H__
+
+#include <iio/iio.h>
+#include "vita49_packet.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* __VITA49_2_BACKEND_H__ */
