--- iiod/vita49_2/vita49_2_udp_receive.c
@@ -1,7 +1,0 @@
-/* SPDX-License-Identifier: LGPL-2.1-or-later */
-/*
- * libiio - Library for interfacing industrial I/O (IIO) devices
- *
- * Copyright (C) 2026 Analog Devices, Inc.
- * Author: Praveen Perera <praveen.perera@analog.com>
- */
